Создал отдельный публичный класс для команд для дальнейшего использования из UI приложения
